PatentSpider/requirements.txt
redis==2.10.5
mrq
requests
lxml
sqlalchemy
psycopg2-binary
simhash
fake_useragent